cleanup origin header
